[feat] #41 코드 리뷰 반영
