+++ app/build.gradle
@@ -86,9 +86,9 @@ dependencies {
     // SwitchButton
     implementation "com.github.zcweng:switch-button:$switch_button_version"
 
-    // fcm
-    implementation "com.google.firebase:firebase-messaging:23.1.1"
-    implementation "com.google.firebase:firebase-bom:31.1.1"
+    // FirebaseCloudMessage
+    implementation "com.google.firebase:firebase-messaging:$fb_messaging_version"
+    implementation "com.google.firebase:firebase-bom:$fb_bom_version"
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.4'
