+++ build.gradle
@@ -15,6 +15,8 @@ buildscript {
         swipe_refresh_layout_version = '1.1.0'
         switch_button_version = "0.0.3@aar"
         fcm_version = "23.1.1"
+        fb_messaging_version = "23.1.1"
+        fb_bom_version = "31.1.1"
     }
     repositories {
         mavenCentral()
